Refactor project to modular MVC and add Mongo model
